Updates: Fix to essencecount and statusbar
Changed area blasts to be fallthrough so that essencecount takes them
into counter as well. Added new draft file lelut (translates as "toys")
which handles all zap items and their cooldowns. Some additions to misc
triggers which are mostly EQ related (planning to split these files to
smaller files in the future). Note that stattusbar requires you to have
a stethoscope for heartbeats!